# Stop looking up the methods at compile time. For much speed.