--- a/Project.toml
+++ b/Project.toml
@@ -14,6 +14,7 @@ Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 CodeTracking = "0.4.0"
 OrderedCollections = "1.1"
 Revise = "2"
+Cassette = "0.2.2"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"