--- a/README.md
+++ b/README.md
@@ -21,8 +21,9 @@ Within this you can read (and write) variables,
 
  - `set_breakpoint!([function|method])`: Set a breakpoint on call to the argument
  - `set_breakpoint!(filename, line number)`: Set a breakpoint on the given line in the given function
- - `set_nodebug!([function|method|module])`: Disable debugging in the given function/method/module
-    - Not having debugging enabled for modules that are not between you and your breakpoints massively speeds up the running of your program.
+ - `set_nodebug!([function|module])`: Disable debugging in the given function/module
+    - Not having debugging enabled for modules/functions you do not need to debug massively speeds up the running of your program.
+    - However, debugging is fully disabled for those modules/functions, so if those functions would then call functions you do want to debug (say by using `map`) then that will also not be caught by the debugger.
  - `list_breakpoints()`, `list_nodebugs()`: list all the breakpoints/nodebugs
  - `rm_breakpoint!(arg...)`, `rm_nodebug!(args...)`: remove breakpoints/nodebugs. Takes same arguments as `set_...`.
  - `clear_breakpoints!()`, `clear_nodebugs!()`: remove all breakpoints/nodebugs.