--- a/src/MagneticReadHead.jl
+++ b/src/MagneticReadHead.jl
@@ -9,7 +9,7 @@ using CodeTracking
 # We don't use Revise, but if it isn't loaded CodeTracking has issues
 using Revise: Revise
 
-export @iron_debug 
+export @iron_debug
 
 include("utils.jl")
 include("method_utils.jl")