--- a/src/core_control.jl
+++ b/src/core_control.jl
@@ -55,10 +55,9 @@ function Cassette.overdub(ctx::HandEvalCtx, @nospecialize(f), @nospecialize(args
     # This is basically the epicenter of all the logic
     # We control the flow of stepping modes
     # and which methods are instrumented or not.
-    method = methodof(f, args...)
     should_recurse =
         ctx.metadata.stepping_mode isa StepIn ||
-        should_instrument(ctx.metadata.breakpoint_rules, method)
+        should_instrument(ctx.metadata.breakpoint_rules, DispatchAttempt(f, args))
 
     if should_recurse
         if Cassette.canrecurse(ctx, f, args...)
@@ -69,7 +68,8 @@ function Cassette.overdub(ctx::HandEvalCtx, @nospecialize(f), @nospecialize(args
                 ctx.metadata.stepping_mode = parent_stepping_mode(_ctx)
             end
         else
-            @warn "Not able to enter into method." f method
+            @assert f isa Core.Builtin
+            #@warn "Not able to enter into method" f args
             return Cassette.fallback(ctx, f, args...)
         end
     else  # !should_recurse