--- a/src/core_control.jl
+++ b/src/core_control.jl
@@ -51,13 +51,13 @@ function Cassette.overdub(::typeof(HandEvalCtx()), args...)
 end
 
 
-function Cassette.overdub(ctx::HandEvalCtx, @nospecialize(f), @nospecialize(args...))
+function Cassette.overdub(ctx::HandEvalCtx, f, @nospecialize(args...))
     # This is basically the epicenter of all the logic
     # We control the flow of stepping modes
     # and which methods are instrumented or not.
     should_recurse =
         ctx.metadata.stepping_mode isa StepIn ||
-        should_instrument(ctx.metadata.breakpoint_rules, DispatchAttempt(f, args))
+        should_instrument(ctx.metadata.breakpoint_rules, f)
 
     if should_recurse
         if Cassette.canrecurse(ctx, f, args...)