@oxinabox 2019-04-22
Ok, I went with disabling setting of nodebug on methods.
Not really a useful feature anyway.

Assuming CI passes, then I will probably merge this tomorrow.